// src/app/main/pages/org-info/models/org-info.interface.ts
export interface IOrgInfo {
    companyCode: string;
    aliasName?: string;
    makerDate: number;
    code: string;
    authStatus: string;
    templateVersionId: number;
    description: string;
    templateTypeName: string;
    version: number;
    extValue?: any;
    templateTypeCode: string;
    modNo: number;
    templateTypeId: number;
    companyId: number;
    recordStatus: string;
    checkerDate?: number;
    name: string;
    checkerId?: string;
    id: number;
    ftsValue?: any;
    effectiveDate: number;
    makerId: string;
}